requirements.txt
streamlit==1.31.0
pydantic==2.6.1
pyyaml==6.0.1
pandas==2.2.0
pytest==8.0.0 